requirements.txt
PyGithub   
slackclient
